System.Process.Common

Documentation

class ProcessMaker a where

Methods

process :: a -> IO (Handle, Handle, Handle, ProcessHandle)

showProcessMakerForUser :: a -> String

class ListLikeIO text char => ListLikeProcessIO text char where

Process IO is based on the ListLikeIO class from the ListLike package

Methods

forceOutput :: text -> IO text

readChunks :: Handle -> IO [text]

Read from a handle, returning a lazy list of the monoid a.

class (IsString text, Monoid text, ListLike text char) => ProcessText text char

class Monoid result => ProcessResult text result | result -> text where

Methods

pidf :: ProcessHandle -> result

outf :: text -> result

errf :: text -> result

intf :: SomeException -> result

codef :: ExitCode -> result

readProcessWithExitCode

Arguments

:: ListLikeProcessIO text char

=> FilePath
command to run

-> [String]
any arguments

-> text
standard input

-> IO (ExitCode, text, text)
exitcode, stdout, stderr

Like readProcessWithExitCode, but with generalized input and output type. Aside from the usual text-like types, the output can be a list of Chunk a. This lets you process the chunks received from stdout and stderr lazil, in the order they are received, as well as the exit code. Utilities to handle Chunks are provided in System.Process.ListLike.

readCreateProcessWithExitCode

Arguments

:: (ProcessMaker maker, ListLikeProcessIO text char)

=> maker
command and arguments to run

-> text
standard input

-> IO (ExitCode, text, text)
exitcode, stdout, stderr

readCreateProcessStrict :: (ProcessMaker maker, ProcessResult text result, ListLikeProcessIO text char) => maker -> text -> IO result

readCreateProcessLazy :: (ProcessMaker maker, ProcessResult a b, ListLikeProcessIO a c) => maker -> a -> IO b

Like readCreateProcessStrict, but the output is read lazily.

showCmdSpecForUser :: CmdSpec -> String

showCreateProcessForUser :: CreateProcess -> String

System.Process utility functions.

Orphan instances

Semigroup ExitCode

Instance details

Methods

(<>) :: ExitCode -> ExitCode -> ExitCode

sconcat :: NonEmpty ExitCode -> ExitCode

stimes :: Integral b => b -> ExitCode -> ExitCode

Monoid ExitCode
A process usually has one ExitCode at the end of its output, this Monoid instance lets us build the type returned by readProcessWithExitCode.

Instance details

Methods

mempty :: ExitCode

mappend :: ExitCode -> ExitCode -> ExitCode

mconcat :: [ExitCode] -> ExitCode